# Docs/openapi.yaml
openapi: 3.0.3
info:
  title: tracklist API
  description: |
    REST API для автентифікації, керування користувачами,
    медіа-контентом, відгуками, коментарями, списками та взаємодіями.
  version: 1.0.0
servers:
  - url: http://localhost:7288
    description: Local ASP.NET Core API
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Логін
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: user@mail.com
              password: password123
      responses:
        "200":
          description: Successfully logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              example:
                data:
                  email: user@mail.com
                  token: jwt-token
        "400":
          description: Wrong login or password
  /api/auth/renewToken:
    post:
      tags: [Auth]
      summary: Оновлення JWT токену
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenewTokenRequest'
            example:
              refreshToken: refresh-token
      responses:
        "200":
          description: Token renewed
          content:
            application/json:
              example:
                data:
                  token: new-jwt-token
        "400":
          description: JWT expired or invalid

  # ===================== USERS =====================
...
  /api/media/translations/{id}:
    put:
      tags: [Media]
      summary: Редагування перекладу медіа
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaTranslationRequest'
      responses:
        "200":
          description: Translation updated

    delete:
      tags: [Media]
      summary: Видалення перекладу медіа
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Translation deleted
...
components:
  schemas:
    UserDto:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        role:
          type: string
        profilePic:
          type: string
          nullable: true
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    RenewTokenRequest:
      type: object
      properties:
        refreshToken:
          type: string
    UserResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/UserDto'
...
    MediaTranslationRequest:
      type: object
      properties:
        languageCode:
          type: string
        title:
          type: string
        description:
          type: string
...